% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf.R
\name{is_udf}
\alias{is_udf}
\title{Test whether an object is a UDF}
\usage{
is_udf(x)
}
\arguments{
\item{x}{An object}
}
\description{
\code{is_udf} returns \code{TRUE} if the given object is a UDF object or subclass thereof
}
\examples{
\dontrun{
u1 <- read_udf("test_udf.yml")
is_udf(u1)
}
}
