% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udfparameter.R
\name{is_udfparameter}
\alias{is_udfparameter}
\title{Test whether an object is a UDF parameter}
\usage{
is_udfparameter(x)
}
\arguments{
\item{x}{An object}
}
\description{
\code{is_udfparameter} returns \code{TRUE} if the given object is a \code{udfparameter} object or subclass thereof
}
