% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_udf.R
\name{read_udf}
\alias{read_udf}
\title{Read a UDF definition file}
\usage{
read_udf(file, ...)
}
\arguments{
\item{file}{either a character string naming a file or a \link{connection}
    open for writing}

\item{...}{
  Arguments passed on to \code{\link[yaml:read_yaml]{yaml::read_yaml}}
  \describe{
    \item{\code{fileEncoding}}{character string: if non-empty declares the
    encoding used on a file (not a connection) so the character data can
    be re-encoded.  See \code{\link{file}}.}
    \item{\code{text}}{character string: if \code{file} is not supplied and this is,
    then data are read from the value of \code{text} via a text connection.
    Notice that a literal string can be used to include (small) data sets
    within R code.}
    \item{\code{error.label}}{a label to prepend to error messages (see Details).}
  }}
}
\value{
A \code{\link{udf}} object representing the UDF
}
\description{
\code{read_udf()} loads a UDF definition from a file
}
\examples{
\dontrun{
my_udf <- read_udf("my_udf.yaml")
}
}
