% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf.R
\name{udf}
\alias{udf}
\title{User-Defined Function (UDF) object}
\description{
User-Defined Function (UDF) object

User-Defined Function (UDF) object
}
\details{
UDF objects store the definition of a user-defined function along with
metadata and unit tests.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{udf$new()}}
\item \href{#method-write}{\code{udf$write()}}
\item \href{#method-create}{\code{udf$create()}}
\item \href{#method-drop}{\code{udf$drop()}}
\item \href{#method-test}{\code{udf$test()}}
\item \href{#method-clone}{\code{udf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{udf} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udf$new(
  name,
  version,
  returns,
  volatility = c("volatile", "stable", "immutable"),
  language = c("sql", "plpythonu"),
  params = list(),
  body,
  description,
  schema = NA,
  replace = FALSE,
  authors = list(),
  tests = list(),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The UDF's name (e.g., \code{"f_calculate_stuff"})}

\item{\code{version}}{A version string (e.g., \code{"1.0"})}

\item{\code{returns}}{Data type returned by this UDF (e.g., \code{"float"})}

\item{\code{volatility}}{Given the same inputs, the function will always return
the same result (\code{"immutable"}), will return the same result during a
given statement (\code{"stable"}), may return different results on successive
calls (\code{"volatile"}).}

\item{\code{language}}{String indicating the language used for the UDF's body
(e.g., \code{"sql"})}

\item{\code{params}}{A list of zero or more \code{\link{udfparameter}} objects specifying
input parameters}

\item{\code{body}}{The body of the UDF containing SQL, Python, or other
supported language}

\item{\code{description}}{A detailed description of the UDF}

\item{\code{schema}}{Database schema in which to place UDF (defualt: \code{NA})}

\item{\code{replace}}{Logical value indicating whether existing versions should
be replaced (default: \code{FALSE})}

\item{\code{authors}}{List of authors (e.g., \code{list("Jane Doe")})}

\item{\code{tests}}{A list of zero or more \code{\link{udftest}} objects specifying test
cases for this UDF}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write UDF definition file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udf$write(file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to \code{\link[yaml:write_yaml]{yaml::write_yaml()}}}

\item{\code{File}}{File to write to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udf$create(conn, test = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop"></a>}}
\if{latex}{\out{\hypertarget{method-drop}{}}}
\subsection{Method \code{drop()}}{
Drop (delete) UDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udf$drop(conn, cascade = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conn}}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{\code{cascade}}{Whether or not to automatically drop objects that depend
on the UDF (default: \code{FALSE})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test"></a>}}
\if{latex}{\out{\hypertarget{method-test}{}}}
\subsection{Method \code{test()}}{
Run tests TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udf$test(conn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conn}}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
