% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udfparameter.R
\name{udfparameter}
\alias{udfparameter}
\title{User Defined Function (UDF) parameter object}
\description{
\code{udfparameter} objects store details about parameters to user-defined functions.
Each \code{udfparameter} object has the following properties:

\describe{
\item{name}{Parameter's name (e.g., \code{"Sepal.Length"})}
\item{type}{Parameter's data type (e.g., \code{FLOAT})}
\item{description}{Detailed description of the parameter}
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{udfparameter$new()}}
\item \href{#method-clone}{\code{udfparameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{udfparameter} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udfparameter$new(name, type, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The parameter's name (e.g., "Sepal.Length"")}

\item{\code{type}}{The parameter's data type (e.g., "FLOAT")}

\item{\code{description}}{A description of the parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udfparameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
