% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udftest.R
\name{udftest}
\alias{udftest}
\title{User Defined Function (UDF) unit test object}
\description{
\code{udftest} objects store details about unit tests for user-defined functions.
Each \code{udftest} object has the following properties:

\describe{
\item{query}{Query using the UDF. The UDF's name should be represented by '?' so that name changes are automatically handled. For example, \code{"SELECT ?(21)"}, would call the UDF with input value 21.}
\item{expect}{Expected value returned by the query}
\item{description}{Detailed description of the test}
\item{example}{Whether or not to include this test in any generated documentation}
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{udftest$new()}}
\item \href{#method-clone}{\code{udftest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{udftest} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udftest$new(query, expect, description, example = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{Query using the UDF}

\item{\code{expect}}{Expected value returned by the query}

\item{\code{description}}{Detailed description of the test}

\item{\code{example}}{Whether or not to include this test in any generated documentation (default: \code{TRUE})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{udftest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
