% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udftool-package.R
\docType{package}
\name{udftool-package}
\alias{udftool}
\alias{udftool-package}
\title{udftool: Tools for managing PostgreSQL user-defined functions}
\description{
Tools for developing, deploying, and maintaining PostgreSQL
    user-defined functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://briandconnelly.github.io/udftool}
  \item \url{https://github.com/briandconnelly/udftool}
}

}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net} (\href{https://orcid.org/0000-0002-9948-0379}{ORCID})

}
\keyword{internal}
