% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_udf.R
\name{write_udf}
\alias{write_udf}
\title{Write UDF definition file}
\usage{
write_udf(x, file, ...)
}
\arguments{
\item{x}{the object to be converted}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing}

\item{...}{Additional arguments passed to \code{\link[yaml]{write_yaml}}}
}
\description{
Write UDF definition file
}
\examples{
\dontrun{
write_udf(my_udf, "my_udf.yml")
}
}
