% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_udf.R
\name{drop_udf}
\alias{drop_udf}
\title{Drop a UDF}
\usage{
drop_udf(udf, conn, cascade = FALSE)
}
\arguments{
\item{udf}{A \code{\link{udf}} object}

\item{conn}{A \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}} object, as returned by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{cascade}{Whether to automatically drop objects that depend on the
function (default: \code{FALSE})}
}
\value{
\code{TRUE}, invisibly, if successful
}
\description{
\code{drop_udf()} deletes a UDF from the given database
}
\examples{
\dontrun{
TODO
}
}
