% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_udf.R
\name{load_udf}
\alias{load_udf}
\title{Load (i.e., create) a user-defined function}
\usage{
load_udf(udf, conn, test = TRUE)
}
\arguments{
\item{udf}{A \code{\link{udf}} object}

\item{conn}{A \link[DBI]{DBIConnection-class} object, as returned by \link[DBI]{dbConnect}}

\item{test}{Whether or not to run tests after loading (default: \code{TRUE})}
}
\description{
\code{load_udf} execucutes a \code{CREATE FUNCTION} statement to load a user-defined function on a given database server.
}
\examples{
\dontrun{
my_udf <- read_udf("my_udf.yaml")
my_db_connection <- DBI::dbConnect()
load_udf(my_udf, my_db_connection)
}
}
