% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_udf.R
\name{read_udf}
\alias{read_udf}
\title{Read a UDF definition file}
\usage{
read_udf(file, ...)
}
\arguments{
\item{file}{either a character string naming a file or a \link{connection}
    open for writing}

\item{...}{Additional arguments passed to \code{\link[yaml]{read_yaml}}}
}
\value{
A \code{\link{udf}} object representing the UDF
}
\description{
\code{read_udf} loads a UDF definition from a file. For more information
on the udf object returned, see \code{\link{udf}}.
}
\examples{
\dontrun{
my_udf <- read_udf("my_udf.yaml")
}
}
