% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_udf.R
\name{test_udf}
\alias{test_udf}
\title{Run a UDF's tests}
\usage{
test_udf(udf, conn)
}
\arguments{
\item{udf}{A \code{\link{udf}} object}

\item{conn}{A \link[DBI]{DBIConnection-class} object, as returned by \link[DBI]{dbConnect}}
}
\description{
\code{test_udf} TODO
}
\examples{
\dontrun{
TODO
}
}
