% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_udf.R
\name{write_udf}
\alias{write_udf}
\title{Write UDF definition file}
\usage{
write_udf(x, file, ...)
}
\arguments{
\item{x}{A \code{\link{udf}} object}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing}

\item{...}{
  Arguments passed on to \code{\link[yaml:write_yaml]{yaml::write_yaml}}
  \describe{
    \item{\code{fileEncoding}}{character string: if non-empty declares the
    encoding to be used on a file (not a connection) so the character data can
    be re-encoded as they are written.  See \code{\link{file}}.}
  }}
}
\description{
\code{write_udf()} writes a \code{\link{udf}} object to a file
}
\examples{
\dontrun{
write_udf(my_udf, "my_udf.yml")
}
}
